+++ src/main/java/frc/robot/subsystems/Intake/CoralIntakeIOSim.java
@@ -2,4 +2,17 @@ package frc.robot.subsystems.Intake;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
-public class CoralIntakeIOSim extends SubsystemBase {}
+public class CoralIntakeIOSim implements CoralIntakeIO {
+  public void moveLeftMotor(double Speed) {
+  }
+
+  public void moveRightMotor(double Speed) {
+  }
+
+  public boolean checkSensor() {
+    return false;
+  }
+
+  public void stop() {
+  }
+}
